Ticket #5831: Raise exception if screenshot was not taken

+++ app/workers/screenshot_worker.rb
@@ -16,8 +16,9 @@ class ScreenshotWorker
     output_file = File.join(Rails.root, 'public', 'screenshots', tmp)
     begin
       fetcher.take_screenshot! url: url, output: output_file, wait_for_element: ['body'], sleep: 10, frames_path: []
+      raise 'Could not take screenshot' unless File.exists?(output_file)
       FileUtils.rm_f path
-      FileUtils.mv output_file, path
+      FileUtils.mv(output_file, path)
     rescue Exception => e
       FileUtils.rm_f path
       FileUtils.ln_s File.join(Rails.root, 'public', 'no_picture.png'), path
